pyproject.toml
[project]
name = "smolagent"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "GTimothee", email = "ava_punksmash@protonmail.com"},
]
dependencies = ["smolagents>=1.13.0", "python-dotenv>=1.1.0", "gradio>=5.25.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
